+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/jsr305/TypeQualifierResolver.java
@@ -66,6 +66,8 @@ public class TypeQualifierResolver {
         DescriptorFactory.createClassDescriptor("org/checkerframework/checker/nullness/compatqual/NonNullDecl"),
         DescriptorFactory.createClassDescriptor("org/netbeans/api/annotations/common/NonNull"),
         DescriptorFactory.createClassDescriptor("lombok/NonNull"),
+        DescriptorFactory.createClassDescriptor("org/jspecify/annotations/NonNull"),
+        DescriptorFactory.createClassDescriptor("org/jspecify/annotations/NullMarked"),
     };
 
     private static final ClassDescriptor[] NULLABLE_CLASS_DESCRIPTORS = new ClassDescriptor[] {
@@ -81,6 +83,8 @@ public class TypeQualifierResolver {
         DescriptorFactory.createClassDescriptor("org/checkerframework/checker/nullness/qual/Nullable"),
         DescriptorFactory.createClassDescriptor("org/checkerframework/checker/nullness/compatqual/NullableDecl"),
         DescriptorFactory.createClassDescriptor("org/netbeans/api/annotations/common/NullAllowed"),
+        DescriptorFactory.createClassDescriptor("org/jspecify/annotations/Nullable"),
+
     };
 
     /**